Split python project into multiple files to aid compilation and maintenance
