+++ wrappers/python/rs_internal.cpp
@@ -1,0 +1,12 @@
+/* License: Apache 2.0. See LICENSE file in root directory.
+Copyright(c) 2017 Intel Corporation. All Rights Reserved. */
+
+#include "python.hpp"
+#include "../include/librealsense2/hpp/rs_internal.hpp"
+
+void init_internal(py::module &m) {
+    /** rs_internal.hpp **/
+    // rs2::software_sensor
+    // rs2::software_device
+    /** end rs_internal.hpp **/
+}
